refactor: 당첨 통계 관련 로직: Lottos -> LottoRank 역할 변경, 클래스명 수정
