+++ src/main/java/step2/view/OutputView.java
@@ -1,9 +1,8 @@
 package step2.view;
 
 import step2.domain.entity.Lottos;
+import step2.domain.entity.LottoRank;
 import step2.domain.vo.LottoPrize;
-
-import java.util.List;
 
 public class OutputView {
 
@@ -18,16 +17,16 @@ public class OutputView {
         lottos.getLottos().forEach(System.out::println);
     }
 
-    public static void printWinnerStatistics(Lottos lottos, List<LottoPrize> lottoPrizes) {
+    public static void printWinnerStatistics(Lottos lottos, LottoRank lottoRank) {
         System.out.println(System.lineSeparator() + "당첨 통계");
         System.out.println("---------");
 
-        printPrizeCount(lottoPrizes);
+        printPrizeCount(lottoRank);
 
-        System.out.println("총 수익률은 " + lottos.getTotalRate(lottoPrizes) + "입니다.");
+        System.out.println("총 수익률은 " + lottoRank.getTotalRate(lottos.size()) + "입니다.");
     }
 
-    private static void printPrizeCount(List<LottoPrize> lottoPrizes) {
+    private static void printPrizeCount(LottoRank lottoRank) {
         final var prizes = LottoPrize.values();
         final var fourthIndex = prizes.length - 2;
         final var firstIndex = 0;
@@ -35,7 +34,7 @@ public class OutputView {
         // 낙첨을 제외한 3개 일치(4등) ~ 6개 일치(1등) 까지 출력
         for (int i = fourthIndex; i >= firstIndex; i--) {
             final var prize = prizes[i];
-            final var count = Lottos.getLottoRankCount(prize, lottoPrizes);
+            final var count = lottoRank.getLottoRankCount(prize);
             System.out.println(prize.getDescription() + " - " + count + "개");
         }
     }
